requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
Werkzeug==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
